requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.6
bcrypt==3.2.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
Faker==37.5.3
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
license-expression==30.4.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
orjson==3.11.1
packageurl-python==0.17.5
packaging==25.0
passlib==1.7.4
pbr==6.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
py-serializable==2.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
ruff==0.12.8
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
starlette==0.47.2
stevedore==5.4.1
toml==0.10.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
